build: say every windows warning at once
/WX stops at the first file with one, which is a round trip of the build for
each. The warnings are collected and the build refuses on any of them at the
end, so one round says all of them.

+++ src/fmt.c
@@ -444,7 +444,7 @@ static void print_operand(Printer *printer, const KestExpr *expr, int limit) {
 // How much room ` else -> ...` wants, so an `if` that gives a value knows
 // whether the rest of it fits on the line it is on.
 static uint32_t else_width(Printer *printer, const KestBranch *branch) {
-    uint32_t width = strlen(" else");
+    uint32_t width = (uint32_t)strlen(" else");
     if (branch->otherwise != NULL) {
         return width + 1 + measure(printer, branch->otherwise);
     }
